+++ src/Tests/AllTests.fs
@@ -43,6 +43,7 @@ let allTests =
                 Maps.tests
                 NetworkSecurityGroup.tests
                 PostgreSQL.tests
+                ResourceGroup.tests
                 ServiceBus.tests
                 SignalR.tests
                 Sql.tests